Plugins: categorize direct_system_calls as a malware plugin
